correcting format specifier for double floating-point numbers

+++ add2Doubles.c
@@ -7,15 +7,15 @@
 #include <stdio.h>
 
 // declare a function that gets an integer from the user
-int getDouble(void) {
-  float input;
-  scanf("%f", &input);
+double getDouble(void) {
+  double input;
+  scanf("%lf", &input);
   return input;
 }
 
 
 int main(int argc, char *argv[]) {
-  float n1, n2, total;
+  double n1, n2, total;
 
   printf("This program adds two numbers.\n");
   printf("First number? ");
@@ -23,7 +23,7 @@ int main(int argc, char *argv[]) {
   printf("Second number? ");
   n2 = getDouble();
   total = n1 + n2;
-  printf("Your total is %f.\n", total);
+  printf("Your total is %lg.\n", total);
 
   return 0;
 }
